+++ src/main/java/gtPlusPlus/xmod/bartcrops/abstracts/BaseCrop.java
@@ -25,7 +25,7 @@ public abstract class BaseCrop extends CropCard implements ICropCardInfo {
     }
 
     public float dropGainChance() {
-        return (float) (Math.pow(0.95D, (double) ((float) this.tier())) * (double) 1f);
+        return (float) (Math.pow(0.95D, (float) this.tier()) * (double) 1f);
     }
 
     public boolean canCross(ICropTile crop) {
